refactor: centralize sanitizeInput util
