+++ html/contact-center.html
@@ -21,6 +21,6 @@
     <footer>
         <p>&copy; 2025 Ops Online Support. All rights reserved.</p>
     </footer>
-    <script src="../js/pages/main.js" defer></script>
+    <script type="module" src="../js/pages/main.js" defer></script>
 </body>
 </html>
