+++ html/modals/contact_us_modal.html
@@ -106,6 +106,6 @@
   </div>
 </div>
 <!-- Link to the JavaScript file that will control the modal's behavior -->
-<script src="../js/pages/contact_us.js"></script>
+<script type="module" src="../js/pages/contact_us.js" defer></script>
 </body>
 </html>
