+++ index.html
@@ -182,8 +182,8 @@
   <!-- ==================================================================
        JavaScript
        ================================================================== -->
-  <script src="js/pages/main.js" defer></script>
-  <script src="js/pages/contact_us.js" defer></script>
+  <script type="module" src="js/pages/main.js" defer></script>
+  <script type="module" src="js/pages/contact_us.js" defer></script>
   <script src="js/pages/join_us.js" defer></script>
   <script src="js/pages/chatbot.js" defer></script>
 </body>
